Add login security to profiles and projects views

+++ projects/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Project, Category, Country
@@ -76,8 +77,13 @@ def project_description(request, project_id):
     return render(request, 'projects/project_description.html', context)
 
 
+@login_required
 def add_project(request):
     """ Add a new project """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProjectForm(request.POST, request.FILES)
         if form.is_valid():
@@ -98,6 +104,7 @@ def add_project(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_project(request, project_id):
     """ Edit an existing project """
 
@@ -128,8 +135,13 @@ def edit_project(request, project_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_project(request, project_id):
     """ Delete an existing project """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
 
     project = get_object_or_404(Project, pk=project_id)
     project.delete()
